# docs/swagger.yaml
definitions:
  db.CardModel:
    properties:
      authToken:
        type: string
      cardType:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      expiryMonth:
        type: string
      expiryYear:
        type: string
      id:
        type: string
      isApproved:
        type: boolean
      isDefault:
        type: boolean
      lastDigits:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/db.UserModel'
      userId:
        type: string
    required:
    - authToken
    - cardType
    - createdAt
    - deletedAt
    - expiryMonth
    - expiryYear
    - id
    - isApproved
    - isDefault
    - lastDigits
    - updatedAt
    - user
    - userId
    type: object
  db.Currency:
    enum:
    - NGN
    type: string
    x-enum-varnames:
    - CurrencyNgn
  db.DeploymentLogModel:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      deployment:
        $ref: '#/definitions/db.DeploymentModel'
      deploymentId:
        type: string
      id:
        type: string
      jobId:
        type: string
      message:
        type: string
      updatedAt:
        type: string
    required:
    - createdAt
    - deletedAt
    - deployment
    - deploymentId
    - id
    - jobId
    - message
    - updatedAt
    type: object
  db.DeploymentModel:
    properties:
      DeploymentLog:
        items:
          $ref: '#/definitions/db.DeploymentLogModel'
        type: array
      actor:
        type: string
      commitHash:
        type: string
      commitMessage:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: string
      machine:
        $ref: '#/definitions/db.MachineModel'
      machineId:
        type: string
      repoConnection:
        $ref: '#/definitions/db.RepoConnectionModel'
      repoConnectionId:
        type: string
      status:
        $ref: '#/definitions/db.DeploymentStatus'
      updatedAt:
        type: string
    required:
    - DeploymentLog
    - actor
    - commitHash
    - commitMessage
    - createdAt
    - deletedAt
    - id
    - machine
    - machineId
    - repoConnection
    - repoConnectionId
    - status
    - updatedAt
    type: object
  db.DeploymentStatus:
    enum:
    - IDLE
    - IN_PROGRESS
    - SUCCESSFUL
    - FAILED
    type: string
    x-enum-varnames:
    - DeploymentStatusIdle
    - DeploymentStatusInProgress
    - DeploymentStatusSuccessful
    - DeploymentStatusFailed
  db.GithubAccountConnectionModel:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      githubEmail:
        type: string
      githubId:
        type: string
      githubInstallationId:
        type: integer
      githubUsername:
        type: string
      id:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/db.UserModel'
      userId:
        type: string
    required:
    - createdAt
    - deletedAt
    - githubEmail
    - githubId
    - githubInstallationId
    - githubUsername
    - id
    - updatedAt
    - user
    - userId
    type: object
  db.InvoiceModel:
    properties:
      createdAt:
        type: string
      currency:
        $ref: '#/definitions/db.Currency'
      deletedAt:
        type: string
      description:
        type: string
      from:
        type: string
      id:
        type: string
      productId:
        type: string
      quantity:
        type: integer
      reference:
        type: string
      status:
        $ref: '#/definitions/db.InvoiceStatus'
      to:
        type: string
      totalPrice:
        type: number
      unitPrice:
        type: number
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/db.UserModel'
      userId:
        type: string
    required:
    - createdAt
    - currency
    - deletedAt
    - description
    - from
    - id
    - productId
    - quantity
    - reference
    - status
    - to
    - totalPrice
    - unitPrice
    - updatedAt
    - user
    - userId
    type: object
  db.InvoiceStatus:
    enum:
    - PAID
    - UNPAID
    type: string
    x-enum-varnames:
    - InvoiceStatusPaid
    - InvoiceStatusUnpaid
  db.MachineModel:
    properties:
      Deployment:
        items:
          $ref: '#/definitions/db.DeploymentModel'
        type: array
      Network:
        items:
          $ref: '#/definitions/db.NetworkModel'
        type: array
      RepoConnection:
        items:
          $ref: '#/definitions/db.RepoConnectionModel'
        type: array
      containerId:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: string
      machineImage:
        type: string
      machineName:
        type: string
      machineStatus:
        $ref: '#/definitions/db.MachineStatus'
      owner:
        $ref: '#/definitions/db.UserModel'
      ownerId:
        type: string
      plan:
        $ref: '#/definitions/db.MachinePlanModel'
      planId:
        type: string
      updatedAt:
        type: string
    required:
    - Deployment
    - Network
    - RepoConnection
    - containerId
    - createdAt
    - deletedAt
    - id
    - machineImage
    - machineName
    - machineStatus
    - owner
    - ownerId
    - plan
    - planId
    - updatedAt
    type: object
  db.MachinePlanModel:
    properties:
      Machine:
        items:
          $ref: '#/definitions/db.MachineModel'
        type: array
      cpu:
        type: string
      createdAt:
        type: string
      currency:
        $ref: '#/definitions/db.Currency'
      deletedAt:
        type: string
      hourlyRate:
        type: number
      id:
        type: string
      memory:
        type: string
      monthlyRate:
        type: number
      name:
        type: string
      updatedAt:
        type: string
    required:
    - Machine
    - cpu
    - createdAt
    - currency
    - deletedAt
    - hourlyRate
    - id
    - memory
    - monthlyRate
    - name
    - updatedAt
    type: object
  db.MachineStatus:
    enum:
    - CREATING
    - RUNNING
    - REBOOTING
    - SHUTTING_DOWN
    - SHUTDOWN
    type: string
    x-enum-varnames:
    - MachineStatusCreating
    - MachineStatusRunning
    - MachineStatusRebooting
    - MachineStatusShuttingDown
    - MachineStatusShutdown
  db.NetworkModel:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      destinationPort:
        type: integer
      hostName:
        type: string
      id:
        type: string
      ingressId:
        type: string
      listeningPort:
        type: integer
      machine:
        $ref: '#/definitions/db.MachineModel'
      machineId:
        type: string
      protocol:
        type: string
      serviceId:
        type: string
      updatedAt:
        type: string
    required:
    - createdAt
    - deletedAt
    - destinationPort
    - hostName
    - id
    - ingressId
    - listeningPort
    - machine
    - machineId
    - protocol
    - serviceId
    - updatedAt
    type: object
  db.RepoConnectionModel:
    properties:
      Deployment:
        items:
          $ref: '#/definitions/db.DeploymentModel'
        type: array
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: string
      machine:
        $ref: '#/definitions/db.MachineModel'
      machineId:
        type: string
      repoId:
        type: string
      repoName:
        type: string
      updatedAt:
        type: string
    required:
    - Deployment
    - createdAt
    - deletedAt
    - id
    - machine
    - machineId
    - repoId
    - repoName
    - updatedAt
    type: object
  db.SocialConnectionModel:
    properties:
      connectionId:
        type: string
      connectionType:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/db.UserModel'
      userId:
        type: string
    required:
    - connectionId
    - connectionType
    - createdAt
    - deletedAt
    - id
    - updatedAt
    - user
    - userId
    type: object
  db.TransactionModel:
    properties:
      amount:
        type: number
      createdAt:
        type: string
      currency:
        $ref: '#/definitions/db.Currency'
      deletedAt:
        type: string
      description:
        type: string
      id:
        type: string
      reference:
        type: string
      status:
        $ref: '#/definitions/db.TransactionStatus'
      type:
        $ref: '#/definitions/db.TransactionType'
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/db.UserModel'
      userId:
        type: string
    required:
    - amount
    - createdAt
    - currency
    - deletedAt
    - description
    - id
    - reference
    - status
    - type
    - updatedAt
    - user
    - userId
    type: object
  db.TransactionStatus:
    enum:
    - PENDING
    - SUCCESS
    - FAILED
    type: string
    x-enum-varnames:
    - TransactionStatusPending
    - TransactionStatusSuccess
    - TransactionStatusFailed
  db.TransactionType:
    enum:
    - DEBIT
    - CREDIT
    type: string
    x-enum-varnames:
    - TransactionTypeDebit
    - TransactionTypeCredit
  db.UserModel:
    properties:
      Card:
        items:
          $ref: '#/definitions/db.CardModel'
        type: array
      GithubAccountConnection:
        items:
          $ref: '#/definitions/db.GithubAccountConnectionModel'
        type: array
      Invoice:
        items:
          $ref: '#/definitions/db.InvoiceModel'
        type: array
      Machine:
        items:
          $ref: '#/definitions/db.MachineModel'
        type: array
      SocialConnection:
        items:
          $ref: '#/definitions/db.SocialConnectionModel'
        type: array
      Transaction:
        items:
          $ref: '#/definitions/db.TransactionModel'
        type: array
      createdAt:
        type: string
      deletedAt:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: string
      lastName:
        type: string
      password:
        type: string
      roles:
        items:
          type: string
        type: array
      updatedAt:
        type: string
    required:
    - Card
    - GithubAccountConnection
    - Invoice
    - Machine
    - SocialConnection
    - Transaction
    - createdAt
    - deletedAt
    - email
    - firstName
    - id
    - lastName
    - password
    - roles
    - updatedAt
    type: object
  handlers.RedirectResult:
    properties:
      link:
        type: string
    required:
    - link
    type: object
...
  services.DeployRepoArgs:
    properties:
      connectionId:
        type: string
      ref:
        type: string
    required:
    - connectionId
    - ref
    type: object
...
  types.AuthorizeCardArgs:
    properties:
      otp:
        type: string
      reference:
        type: string
    required:
    - otp
    - reference
    type: object
...
  types.ListInvoiceArgs:
    properties:
      search:
        type: string
      skip:
        type: integer
      sortBy:
        type: string
      take:
        type: integer
      userId:
        type: string
    type: object
...
  types.PreAuthorizeCardArgs:
    properties:
      cardNumber:
        type: string
      cvv:
        type: string
      expiryMonth:
        type: string
      expiryYear:
        type: string
      pin:
        type: string
    required:
    - cardNumber
    - cvv
    - expiryMonth
    - expiryYear
    - pin
    type: object
  types.PreAuthorizeCardResult:
    properties:
      reference:
        type: string
    required:
    - reference
    type: object
...
  types.Repository:
    properties:
      fullName:
        type: string
      id:
        type: integer
      name:
        type: string
      private:
        type: boolean
      url:
        type: string
    required:
    - fullName
    - id
    - name
    - private
    - url
    type: object
  types.UpdateCardArgs:
    properties:
      AuthToken:
        type: string
      CardType:
        type: string
      ExpiryMonth:
        type: string
      ExpiryYear:
        type: string
      IsApproved:
        type: boolean
      IsDefault:
        type: boolean
      LastDigits:
        type: string
    type: object
...
info:
  contact:
    email: formatio@overal-x.org
    name: Formatio Team
  title: Formatio API
  version: "1.0"
paths:
...
  /billing/cards:
    get:
      consumes:
      - application/json
      operationId: listCards
      parameters:
      - in: query
        name: search
        type: string
      - in: query
        name: skip
        type: integer
      - in: query
        name: sortBy
        type: string
      - in: query
        name: take
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.CardModel'
            type: array
      tags:
      - billing
  /billing/cards/{cardId}:
    delete:
      consumes:
      - application/json
      operationId: deleteCard
      parameters:
      - description: Card Id
        in: path
        name: cardId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      tags:
      - billing
    patch:
      consumes:
      - application/json
      operationId: UpdateCard
      parameters:
      - description: Card Id
        in: path
        name: cardId
        required: true
        type: string
      - description: Update Card
        in: body
        name: args
        required: true
        schema:
          $ref: '#/definitions/types.UpdateCardArgs'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/db.CardModel'
      tags:
      - billing
  /billing/cards/authorize:
    post:
      consumes:
      - application/json
      operationId: authorizeCard
      parameters:
      - description: Update Repo Connection
        in: body
        name: args
        required: true
        schema:
          $ref: '#/definitions/types.AuthorizeCardArgs'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/db.CardModel'
      tags:
      - billing
  /billing/cards/pre-authorize:
    post:
      consumes:
      - application/json
      operationId: preAuthorizeCard
      parameters:
      - description: Pre-Authorize Card
        in: body
        name: args
        required: true
        schema:
          $ref: '#/definitions/types.PreAuthorizeCardArgs'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.PreAuthorizeCardResult'
      tags:
      - billing
  /billing/invoice:
    get:
      consumes:
      - application/json
      operationId: listInvoice
      parameters:
      - description: List Invoice
        in: body
        name: args
        required: true
        schema:
          $ref: '#/definitions/types.ListInvoiceArgs'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            items:
              $ref: '#/definitions/db.InvoiceModel'
            type: array
      tags:
      - billing
  /billing/invoice/{invoiceId}:
    get:
      consumes:
      - application/json
      operationId: getInvoice
      parameters:
      - description: Invoice Id
        in: path
        name: invoiceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.InvoiceModel'
      tags:
      - billing
  /deployments:
    get:
      consumes:
      - application/json
      operationId: listDeployments
      parameters:
      - description: Machine Id
        in: query
        name: machineId
        type: string
      - in: query
        name: machineId
        required: true
        type: string
      - in: query
        name: repoConnectionId
        type: string
      - in: query
        name: search
        type: string
      - in: query
        name: skip
        type: integer
      - in: query
        name: sortBy
        type: string
      - in: query
        name: take
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.DeploymentModel'
            type: array
      tags:
      - deployments
  /deployments/{deploymentId}:
    get:
      consumes:
      - application/json
      operationId: getDeploymentById
      parameters:
      - description: Deployment Id
        in: path
        name: deploymentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.DeploymentModel'
      tags:
      - deployments
  /deployments/{deploymentId}/logs:
    get:
      consumes:
      - application/json
      operationId: listDeploymentLogsById
      parameters:
      - description: Deployment Id
        in: path
        name: deploymentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.DeploymentLogModel'
            type: array
      tags:
      - deployments
  /deployments/deploy:
    post:
      consumes:
      - application/json
      operationId: deployRepo
      parameters:
      - description: Deploy Repo
        in: body
        name: args
        required: true
        schema:
          $ref: '#/definitions/services.DeployRepoArgs'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      tags:
      - deployments
...
  /gh/authorize:
    get:
      consumes:
      - application/json
      operationId: authorizeGithubAccount
      parameters:
      - in: query
        name: redirectUrl
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.RedirectResult'
      tags:
      - github
  /gh/repos:
    get:
      consumes:
      - application/json
      operationId: ListRepositories
      parameters:
      - default: 1
        in: query
        name: pageNumber
        type: integer
      - default: 20
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.Repository'
            type: array
      tags:
      - github
...
swagger: "2.0"